ai-app-week3/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-jose
passlib
python-dotenv
bcrypt